Feartures
Goal 1
Fetch a random dog image from the API and display it on the page.

Use the random endpoint of the API for this.

https://dog.ceo/api/breeds/image/random

Goal 2
Add a form element that lets users submit their rating, along with a submit button. This can be as simple as a text field, and as complex as a star rating input element.

Goal 3
Once a rating has been submitted, load a new dog image and reset your form elements.